+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.0.5
+Release Date: May 24, 2020
+
+* Improved: Webhook Handler.
+
+## 1.0.4
+Release Date: May 18, 2020
+
+* Improved: Payment verfication mechanism.
+* Improved: Webhook mechanism.
+* Tweak: It is now possible to use any other payment method to make payment when payment initialized at first through the Razoypay Gateway method.
+
+## 1.0.3
+Release Date: May 12, 2020
+
+* NEW: Added Razorpay Webhook feature.
+* NEW: Added Instant Refund option.
+* Fixed: Email notification is not working properly.
+* Improved: Reduced icon image size to adapt the width properly.
+* Compatibilty with WooCommerce 4.1.
+
+## 1.0.2
+Release Date: April 26, 2020
+
+* Compatibilty with WordPress 5.4 and WooCommerce 4.0.
+
 ## 1.0.1
 Release Date: January 30, 2020